Library documentation

Model structures

Main.HubbardFunctions.OB_Sim — Type

OB_Sim(t::Vector{Float64}, u::Vector{Float64}, μ=0.0, J::Vector{Float64}, P=1, Q=1, svalue=2.0, bond_dim=50, period=0; kwargs...)

Construct a parameter set for a 1D one-band Hubbard model with a fixed number of particles.

Arguments

t: Vector in which element $n$ is the value of the hopping parameter of distance $n$. The first element is the nearest-neighbour hopping.

u: Vector in which element $n$ is the value of the Coulomb interaction with site at distance $n-1$. The first element is the on-site interaction.

J: Vector in which element $n$ is the value of the exchange interaction with site at distance $n$. The first element is the nearest-neighbour exchange.

µ: The chemical potential.

P,Q: The ratio P/Q defines the number of electrons per site, which should be larger than 0 and smaller than 2.

svalue: The Schmidt truncation value, used to truncate in the iDMRG2 algorithm for the computation of the groundstate.

bond_dim: The maximal bond dimension used to initialize the state.

Period: Perform simulations on a helix with circumference Period. Value 0 corresponds to an infinite chain.

Put the optional argument spin=true to perform spin-dependent calculations.

Main.HubbardFunctions.OBC_Sim — Type

OBC_Sim(t::Vector{Float64}, u::Vector{Float64}, μf::Float64, svalue=2.0, bond_dim=50, period=0; mu=true, kwargs...)

Construct a parameter set for a 1D one-band Hubbard model with the number of particles determined by a chemical potential.

Arguments

t: Vector in which element $n$ is the value of the hopping parameter of distance $n$. The first element is the nearest-neighbour hopping.

u: Vector in which element $n$ is the value of the Coulomb interaction with site at distance $n-1$. The first element is the on-site interaction.

µf: The chemical potential, if mu=true. Otherwise, the filling of the system. The chemical potential corresponding to the given filling is determined automatically.

svalue: The Schmidt truncation value, used to truncate in the iDMRG2 algorithm for the computation of the groundstate.

bond_dim: The maximal bond dimension used to initialize the state.

Period: Perform simulations on a helix with circumference Period. Value 0 corresponds to an infinite chain.

Spin-dependent calculations are not yet implemented.

Main.HubbardFunctions.MB_Sim — Type

MB_Sim(t::Matrix{Float64}, u::Matrix{Float64}, J::Matrix{Float64}, U13::Matrix{Float64}, P=1, Q=1, svalue=2.0, bond_dim=50; kwargs...)

Construct a parameter set for a 1D B-band Hubbard model with a fixed number of particles.

Arguments

t: Bx(nB) matrix in which element $(i,j)$ is the hopping parameter from band $i$ to band $j$. The on-site, nearest neighbour, next-to-nearest neighbour... hopping matrices are concatenated horizontally.

u: Bx(nB) matrix in which element $(i,j)$ is the Coulomb repulsion $U_{ij}=U_{iijj}$ between band $i$ and band $j$. The on-site, nearest neighbour, next-to-nearest neighbour... matrices are concatenated horizontally.

J: Bx(nB) matrix in which element $(i,j)$ is the exchange $J_{ij}=U_{ijji}=U_{ijij}$ between band $i$ and band $j$. The on-site, nearest neighbour, next-to-nearest neighbour... matrices are concatenated horizontally. The diagonal terms of the on-site matrix are ignored.

U13: BxB matrix in which element $(i,j)$ is the parameter $U_{ijjj}=U_{jijj}=U_{jjij}=U_{jjji}$ between band $i$ and band $j$. Only on-site. The diagonal terms of the on-site matrix are ignored. This argument is optional.

P,Q: The ratio P/Q defines the number of electrons per site, which should be larger than 0 and smaller than 2.

svalue: The Schmidt truncation value, used to truncate in the iDMRG2 algorithm for the computation of the groundstate.

bond_dim: The maximal bond dimension used to initialize the state.

Put the optional argument 'spin=true' to perform spin-dependent calculations.

U13 inter-site, Uijkk, and Uijkl can be inserted using kwargs.

Use the optional argument name to assign a name to the model. This is used to destinguish between different parameter sets: Wrong results could be loaded or overwritten if not used consistently!!!

Main.HubbardFunctions.MBC_Sim — Type

MBC_Sim(t::Matrix{Float64}, u::Matrix{Float64}, J::Matrix{Float64}, U13::Matrix{Float64}, svalue=2.0, bond_dim=50; kwargs...)

Construct a parameter set for a 1D $B$-band Hubbard model with the number of particles determined by a chemical potential.

Arguments

t: $B\times nB$ matrix in which element $(i,j)$ is the hopping parameter from band $i$ to band $j$. The on-site, nearest neighbour, next-to-nearest neighbour... hopping matrices are concatenated horizontally. The diagonal terms of the on-site matrix determine the filling.

u: $B\times nB$ matrix in which element $(i,j)$ is the Coulomb repulsion $U_{ij}=U_{iijj}$ between band $i$ and band $j$. The on-site, nearest neighbour, next-to-nearest neighbour... matrices are concatenated horizontally.

J: $B\times nB$ matrix in which element $(i,j)$ is the exchange $J_{ij}=U_{ijji}=U_{ijij}$ between band $i$ and band $j$. The on-site, nearest neighbour, next-to-nearest neighbour... matrices are concatenated horizontally. The diagonal terms of the on-site matrix are ignored.

U13: $B\times B$ matrix in which element $(i,j)$ is the parameter $U_{ijjj}=U_{jijj}=U_{jjij}=U_{jjji}$ between band $i$ and band $j$. Only on-site. The diagonal terms of the on-site matrix are ignored. This argument is optional.

svalue: The Schmidt truncation value, used to truncate in the iDMRG2 algorithm for the computation of the groundstate.

bond_dim: The maximal bond dimension used to initialize the state.

Spin-dependent calculations are not yet implemented.

U13 inter-site, Uijkk, and Uijkl can be inserted using kwargs.

Use the optional argument name to assign a name to the model. This is used to destinguish between different parameter sets: Wrong results could be loaded or overwritten if not used consistently!!!

Compute states

Main.HubbardFunctions.produce_groundstate — Function

produce_groundstate(model::Simulation; force::Bool=false)

Compute or load groundstate of the model. If force=true, overwrite existing calculation.

Main.HubbardFunctions.produce_excitations — Function

produce_excitations(model::Simulation, momenta, nums::Int64; force::Bool=false, charges::Vector{Float64}=[0,0.0,0], kwargs...)

Compute or load quasiparticle excitations of the desired model.

Arguments

model: Model for which excitations are sought.

momenta: Momenta of the quasiparticle excitations.

nums: Number of excitations.

force: If true, overwrite existing calculation.

charges: Charges of the symmetry sector of the excitations.

Main.HubbardFunctions.produce_bandgap — Function

produce_bandgap(model::Union{OB_Sim, MB_Sim}; resolution::Int64=5, force::Bool=false)

Compute or load the band gap of the desired model.

Main.HubbardFunctions.produce_TruncState — Function

produce_truncstate(model::Simulation, trunc_dim::Int64; trunc_scheme::Int64=0, force::Bool=false)

Compute or load a truncated approximation of the groundstate.

Arguments

model: Model for which the groundstate is to be truncated.

trunc_dim: Maximal bond dimension of the truncated state.

trunc_scheme: Scheme to perform the truncation. 0 = VUMPSSvdCut. 1 = SvdCut.

force: If true, overwrite existing calculation.

Properties of groundstate

Main.HubbardFunctions.dim_state — Function

dim_state(ψ::InfiniteMPS)

Determine the bond dimensions in an infinite MPS.

Main.HubbardFunctions.density_state — Function

density_state(model::Simulation)

Compute the number of electrons per site in the unit cell for the ground state.

Main.HubbardFunctions.density_spin — Function

density_spin(model::Union{OB_Sim,MB_Sim})

Compute the density of spin up and spin down per site in the unit cell for the ground state.

Main.HubbardFunctions.calc_ms — Function

calc_ms(model::Union{OB_Sim,MB_Sim})

Compute the staggered magnetization of the ground state.

Tools

Main.HubbardFunctions.plot_excitations — Function

plot_excitations(momenta, energies; title="Excitation_energies", l_margin=[15mm 0mm])

Plot the obtained energy levels in functions of the momentum.

Main.HubbardFunctions.plot_spin — Function

plot_spin(model::Simulation; title="Spin Density", l_margin=[15mm 0mm])

Plot the spin density of the model throughout the unit cell as a heatmap.

Main.HubbardFunctions.extract_params — Function

extract_params(path::String; range_u::Int64= 1, range_t::Int64=2, range_J::Int64=1, 
                    range_U13::Int64=1, r_1111::Int64 = 1, r_112::Int64 = 1)

Extract the parameters from a params.jl file located at path in PyFoldHub format.